Fix manual and service based currency updates
The background service will always fetch all currencies for the day. It does that 3-4 times a day.

The manual update will fetch the latest currency date from the database use that as an date time offset for the update.
If the latest currency date is older than the current date, then the beginning of the current day will be used as
an offset date for the update.

Phew!