Use explicit context passing and assets loading in APISource
The bug was that a static instance of Properties was being used by the service while the
MainActivity context was already destroyed, i.e., OS killed the activity or the user closed it.

This fix makes APISource and AppAssets take a context parameter instead and the assets are being
loaded on demand afterwards.