Limit getLastRates() rows fetched to 5K
Far from a good solution, but does increase loading times with up to 50%.
A better solution would be required. Perhaps a more robust SQL select query.

A 3 days SQLite database may contain up to 20K rows, so fetching and wrapping
all of that into objects takes a lot of time. A Guava Table was used to cache
the results, but the contains() call still takes up to 30% execution time from
the total getLastRates() execution time.

The current, very basic, solution is to limit the rows fetched to 5K. This is
not a good solution and may lead to some rates, those that were not recently
updated, to not being shown in the CurrenciesFragment. Why 5K? There would be
an average of ~50K rows collected in 3 days in the remote database, or about
4,2K per day.